backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pandas
numpy
joblib
onnxruntime
scikit-learn
shap
fairlearn
imbalanced-learn
celery
PyJWT
aiobotocore